dt5

+++ source/common/benchmarks.c
@@ -53,10 +53,10 @@ void evaluate(Benchmarks* bench, Arguments* args) {
 	sigma = sqrt(sigma - (average * average));
 
 	const double tt = ((double)total_time/1000000000);
-	const int argsTT = (((args->count * args->size) * 8));
+	const long argsTT = (((args->count * args->size) * 8));
 	const double messageRate = (argsTT / (double)tt);
 
-	printf("args %.7d\n", argsTT);
+	printf("args %li\n", argsTT);
 	printf("tt  %.7f\n", tt);
 	printf("messageRate  %.7f\n", messageRate);
